Update virtualmachine.sh mykey generation
I was reading your book and created this script, but I got an error related to mykey it was not found, after searching the documentation I found it's required to be created with the command `aws ec2 create-key-pair --key-name mykey` and to store the content in the file mykey.pem with 400 permissions, and then I could connect to the instance.  I add a comment to explain the command I end up using  `aws ec2 create-key-pair --key-name mykey --query "KeyMaterial" --output text > mykey.pem && chmod 400 mykey.pem`. I hope this could help other people that could face the same problem.